Move code to query builder
